Move log view out of main to accommodate introduction of tabs
